+++ crates/pkgcraft-tools/src/bin/pk/subcmds.rs
@@ -2,15 +2,17 @@ use std::process::ExitCode;
 
 use crate::command::Command;
 
+mod completion;
 mod cpv;
 mod dep;
 mod pkg;
 mod repo;
-mod shellcomp;
 mod version;
 
 #[derive(clap::Subcommand)]
 pub(crate) enum Subcommand {
+    /// Generate shell completion
+    Completion(completion::Command),
     /// Cpv commands
     Cpv(cpv::Command),
     /// Dependency commands
@@ -19,8 +21,6 @@ pub(crate) enum Subcommand {
     Pkg(pkg::Command),
     /// Repository commands
     Repo(repo::Command),
-    /// Generate shell completion
-    ShellComp(shellcomp::Command),
     /// Version commands
     Version(version::Command),
 }
@@ -32,7 +32,7 @@ impl Subcommand {
             Self::Dep(cmd) => cmd.run(),
             Self::Pkg(cmd) => cmd.run(args.load_config()?),
             Self::Repo(cmd) => cmd.run(args.load_config()?),
-            Self::ShellComp(cmd) => cmd.run(),
+            Self::Completion(cmd) => cmd.run(),
             Self::Version(cmd) => cmd.run(),
         }
     }